pyproject.toml
[tool.poetry]
name = "pvconsumer"
authors = ["Open Climate Fix"]
description = "Consumer for PV data from pvoutput.org."
version = "2.1.1"
readme = "README.md"
packages = [{include = "pvconsumer"}]


[tool.poetry.dependencies]
python = "^3.10"
pvoutput-ocf = "0.1.17"
click = "^8.1.3"
requests = "^2.28.2"
pandas = "^1.5.3"
pvsite-datamodel = {git = "https://github.com/openclimatefix/pv-site-datamodel.git", branch = "site-location"}
pydantic = "1.10.5"
sentry-sdk = "^2.1.1"
numpy = "<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
isort = "^5.12.0"
black = "^23.1.0"
testcontainers-postgres = "^0.0.1rc1"
pytest-cov = "^4.0.0"
ruff = "^0.0.252"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
